// src/lib/cardcom/generated/models/transactionReq.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */
import type { AdvancedTran } from './advancedTran';
import type { CardOwnerInformation } from './cardOwnerInformation';
import type { CustomField } from './customField';
import type { DocumentTran } from './documentTran';

export interface TransactionReq {
  /** Cardcom terminal number */
  TerminalNumber: number;
  /**
     * Api Name for authentication
     * @minLength 1
     */
  ApiName: string;
  /** Amount of transaction */
  Amount: number;
  /**
     * credit Card number
     * @nullable
     */
  CardNumber?: string | null;
  /**
     * credit Card token
     * @nullable
     */
  Token?: string | null;
  /**
     * External Merchant Id
     * @nullable
     */
  ExternalMerchantId?: string | null;
  /**
     * credit Card Expiration date
     * @nullable
     */
  CardExpirationMMYY?: string | null;
  /**
     * credit Card CVV
     * @nullable
     */
  CVV2?: string | null;
  /**
     * External Uniq Tran Id , send your uniq trnasaction id to prevent duplication of transaction. if the same ExternalUniqTranId will be send you will receive and error code 608. see 'ExternalUniqUniqTranIdResponse'
     * @nullable
     */
  ExternalUniqTranId?: string | null;
  /**
     * External Uniq Tran Id , send your uniq trnasaction id to prevent duplication of transaction. if the same ExternalUniqTranId will be send you will receive and error code 608. see 'ExternalUniqUniqTranIdResponse'
     * @nullable
     */
  ExternalUniqUniqTranIdResponse?: boolean | null;
  /**
     * Maximum number of payment to show to card holder , default 1
     * @minimum 1
     * @maximum 36
     * @nullable
     */
  NumOfPayments?: number | null;
  CardOwnerInformation?: CardOwnerInformation | null;
  /**
     * ISO Coin Code ,1 - ILS, 2 - USD, else ISO Coin from list : https://en.wikipedia.org/wiki/ISO_4217
     * @nullable
     */
  ISOCoinId?: number | null;
  /**
     * Fill the value of CustomFields information
     * @nullable
     */
  CustomFields?: CustomField[] | null;
  /** Advanced transaction options */
  Advanced?: AdvancedTran | null;
  /** Document information , will create document if transaction succeeded */
  Document?: DocumentTran | null;
}